requirements.txt
flask
sqlalchemy
psycopg2-binary
python-dotenv
alembic
huggingface_hub
openai